requirements.txt
aiogram==3.13.1
SQLAlchemy==2.0.36
greenlet==3.1.1
alembic==1.14.0
aiosqlite==0.20.0
qrcode==7.4.2
pillow==11.0.0
python-dotenv==1.0.1

